First 98 fabonacci starting with 1, 1

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -1,19 +1,19 @@
-#include <stdio.h>
+#incdde <stdio.h>
 /**
  * main-98 fibo numbers
  * Return:0 is success
  */
 int main(void)
 {
-	unsigned long int pre = 1, cur = 2, temp, count, num = 98;
+	unsigned int pre = 1, cur = 2, temp, count, num = 98;
 
-	printf("%lu, %lu, ", pre, cur);
+	printf("%d, %d, ", pre, cur);
 	for (count = 2; count <= num; count++)
 	{
 		temp = cur + pre;
 		pre = cur;
 		cur = temp;
-		printf("%lu, ", cur)
+		printf("%d, ", cur)
 	}
 
 	return (0);
